+++ assets/jsx/workflow-editor/components/data-types/workflow.jsx
@@ -33,6 +33,7 @@ export function WorkflowData() {
                 type: "meta",
                 label: "Metadata",
                 description: "The metadata of the workflow.",
+                table: "_postmeta",
             },
             {
                 name: "execution_id",